requirements.txt
colorama==0.3.9
selenium==3.12.0
tldextract==2.2.0
xvfbwrapper==0.2.9
GitPython==2.1.11
